feat(eth-sender): fix for missing eth_txs_history entries (#2236)
This change introduces an invariant "If the transaction may have been
sent, then the sent_at_block is set". This way we never remove
eth_txs_history entries if the transaction may be included in block. The
downside is that we can't distinguish between not sent transaction and
one that has not been mined, but you can't have it both.
The next step is to remove "send_unsent_transactions" step, but it can't
be done in this PR as there might be transactions in db without
sent_at_block set

---------

Signed-off-by: tomg10 <lemures64@gmail.com>