feat(node_framework): Implement FromContext and IntoContext derive macro (#2330)
## What ❔

- Improves context interfaces: removes unneded `async` when fetching
resources, and made `add_task` accept `T` instead of `Box`.
- Adds previously proposed `FromContext` and `IntoContext` macro that
will allow to redefine the wiring layer interface so that it doesn't
have direct access to the context.

⚠️ I didn't port the whole framework to using macros, since there may be
changes in the macro itself during this review. Once we merge this, I
will simultaneously rework the `WiringLayer` interface and will port
layers. For now, I've used new macros in the `BatchStatusUpdater` layer
to see how it works. Seems that it works fine 😅

## Why ❔

Ergonomics.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.