feat: Base Token Fundamentals (#2204)
## What ❔

This PR puts in place the major pieces needed in the Base Token flow as
described in the [design
doc](https://www.notion.so/matterlabs/Custom-Base-Token-Tech-Design-74e158112f24463c82d96415bed76d04).

This PR adds a preliminary flow of:
- Periodically fetch the new price from an external API
- Update the conversion ration in the DB
- Have the Fee Model use the latest stored conversion ratio

In the scope of this PR:
- Introducing BaseTokenAdjuster to the Node Framework
- Adding the DB migration and the associated DAL
- Adding the proper configs for this layer to work
- Conditionally updating the gas price with a BaseToken<->ETH conversion
ratio such that it's used in charging the gas in the base token and
estimating gas for txs.
- Unit + Integration tests

Out of scope of this PR:
- Protocol contract changes needed for a base token chain to work (these
are already merged)
- Using an external API to fetch prices
- Updating the L1 with the conversion ratio
- Observability, API redundancy,

## Why ❔
Multiple ZK Chains are ready to go on mainnet with a custom token used
as the gas token.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.

---------

Co-authored-by: Ivan Schasny <ivan.schasny@gmail.com>