feat: L1 batch signing (BFT-474) (#2414)
## What ❔

Implements the methods on `PersistentBatchStore` required by consensus
to sign and publish attestations over L1 batches.
Calculates the hash value which needs to be signed by attestors using
`SignedBatchInfo::hash()`.
## Why ❔

This allows consensus to figure out which is the first batch that needs
to be signed after the node starts, and subsequently to retrieve the
payload to be signed as well.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.

---------

Co-authored-by: Bruno França <bruno@franca.xyz>