feat(en): Switch EN to use node framework (#2427)
## What ❔

This PR ports remaining parts of the EN to node framework and makes
framework the default way to run it.
In more detail:

- Config for the health check limits are now set from config.
- EN and rust metrics are now exposed; the protocol version update task
now runs.
- ⚠️ Connection pool healthcheck was removed. It was controversial
initially, its usefulness is not clear, it was supposed to be refactored
a year ago but didn't, and it wasn't working well when testing. See
[linear
issue](https://linear.app/matterlabs/issue/PLA-255/revamp-db-connection-health-check)
for more context.
- Tests were reworked to use node framework; some refactoring was also
applied to reduce boilerplate.
- Additional tests were added to check for invalid EN configurations.
- ⚠️ Node framework was made the default way to run the EN. There is
also a hook to force EN to run the old way, so that we don't have to
rollback over small issues.

## Why ❔

- Part of switch to the node framework.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.