test: Log whether we have attester key (#2437)
## What ❔

Adds debug logs to show whether the new attester key has been parsed.

## Why ❔

On stage2 the main node doesn't seem to run attestations, while the
external node does, even though both have the correct secret keys mapped
to the containers. This is just a sanity check to see whether the values
are present in the parsed configs, it's not an attempt to solve
anything.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.