feat(vlog): New vlog interface + opentelemtry improvements (#2472)
OK, that's a big PR for its purpose, sorry for that.

Essence of changes:

## Nice observability installation from the config

We can now write the following:
```
    let observability_config = configs
        .observability
        .clone()
        .context("observability config")?;
    let _observability_guard = observability_config.install()?;
```

instead of having to manually map data from `observability_config` to
`ObservabilityBuilder` methods.
This is done in a non-intrusive way: the functionality is under a
feature flag in `zksync_config` which is only enabled for binaries.

## Nicer `ObservabilityBuilder` interface

Previously we've had "everything in one place", now we have separate
structures for logs, opentelemetry, and sentry. It would be nice to
integrate prometheus there too, but it's a separate task.

## More service information is exposed via opentelemetry

Previously only service name was exposed. Now we also expose pod name
and namespace.
It's done in a non-intrusive way as well, where you can provide _some_
of information if you have it, but otherwise it'll be populated from env
variables (which is fine, because it's not a permanent configuration,
but rather a machine descriptor).