fix(node): respect namespaces configuration (#2578)
## What ❔

Respect the configuration of namespaces when building the HTTP Web3 API
layer.

## Why ❔

This commit fixes a bug found when trying to run the node server with
the following configuration:
```
API_WEB3_JSON_RPC_API_NAMESPACES=eth,web3,net,pubsub,zks,en,debug,unstable
```
This was added to the `etc/env/target/dev.env` file to test PR #2474. It
turned out that the `unstable` namespace works fine for the WebSocket
API but not for the HTTP API.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.