feat: Poll the main node API for attestation status - relaxed (BFT-496) (#2583)
## What ❔

Reverts https://github.com/matter-labs/zksync-era/pull/2574 to
re-establish the attestation status API integration, but also bumps the
era-consensus version to
[0.1.0-rc.6](https://github.com/matter-labs/era-consensus/pull/174)
which has changes to make it backwards compatible:
* the `Executor` is started without waiting for the initial attestation
batch number to become available, so as not to stop the node completely
from participating in gossip without the main node API (this behaviour
is internal to the `AttesterStatusRunner`)
* invalid batch vote signatures are not rejected, just ignored, so the
node doesn't break connection with a peer who signed a different payload
while the feature is in flux

TODO: 
- [x] Update once `0.1.0-rc.6` is published to crates.io

## Why ❔

The first reason is to not be so eager to drop a gossip peer for
incompatibilities between newer and existing versions of the software
while we're actively working on the features involved.

The second is that the rollout strategy to mainnet seems to be to roll
out external nodes first, main nodes second. It is expected that the
external node should work with an older version of the main node, or at
least not fail to start; it is okay if a new feature doesn't work.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.

---------

Co-authored-by: Bruno França <bruno@franca.xyz>