feat(prover_cli): Add test for status, l1 and config commands. (#2263)
Add tests for the CLI of the proover, for the commands `status l1`,
`status batch`, `and config`.

For now, due to how the configuration setup works, it's necessary to run
the tests sequentially. Eventually, the logic for handling environment
variables can be changed to allow running tests in parallel.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.

---------

Co-authored-by: Joaquin Carletti <joaquin@192.168.1.7>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: Ivan Litteri <ivanlitteri@Ivans-MacBook-Pro.local>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: EmilLuta <EmilLuta@users.noreply.github.com>