fix(prover_cli): Remove congif file check (#2695)
This is a temporary fix to allow the CLI binary to be used directly
without having the repository cloned.

Note: The 'config' command does not work when used this way, and you are
required to specify the database URL each time you use the CLI, as the
config does not persist.

A more reliable solution is being worked on, but in the meantime, it was
necessary to use it this way.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.