chore: Update to latest cargo deny (#2746)
Cargo deny is a tool used for license checks (initial motivation),
vulnerability checks and other checks (such as unmaintained crates).

This tool has been used across core monorepo for a long time, but given
one problematic upgrade it's been pinned to a version. There have been
breaking changes since then and the tool got better, but we got stuck
with the old version. This upgrades to the new version, but is still
pinned. A future development is adding renovate bot, which will keep
version up to date. Currently in backlog of @matter-labs/devops.

This PR updates the tool & addresses issues where the fix is
straightforward. Other issues will need to be prioritized & treated
separately.
They can be found in deny.toml under `advisories.ignore`.

There is space for further improvements on our current defaults, again,
not tackled in this PR.