feat: conditional cbt l1 updates (#2748)
For operational efficiency, this PR enables conditional L1 updates for
chains with custom base token. It adds a new configuration
`BASE_TOKEN_ADJUSTER_L1_UPDATE_DEVIATION` that defines how much the
token price needs to fluctuate in order for the update to be propagated
to L1. Equal to 10% by default.

---------

Co-authored-by: Roman Brodetski <rb@matterlabs.dev>