fix(tee-prover): fix deserialization of `std::time::Duration` in `envy` config (#2817)
## What ❔

This PR fixes the deserialization issue of `std::time::Duration` in the
`envy` configuration.

Relevant logs from the `stage` environment showcasing the issue:
https://grafana.matterlabs.dev/goto/IC-9k4eIR?orgId=1

Error message from the above logs:
```
Error: missing value for field initial_retry_backoff
```

The root cause of the problem supposedly boils down to the mismatch
between the expected format of `TEE_PROVER_INITIAL_RETRY_BACKOFF` and
the actual format. We export it as follows:
```
export TEE_PROVER_INITIAL_RETRY_BACKOFF=1
```
which is not supported as explained here:
https://github.com/serde-rs/serde/issues/339#issuecomment-539453327

## Why ❔

To fix the bug.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.