fix: ignore unknown fields in rpc json response (#2962)
In some previous pr, I've added a field to json RPC response
(consensusGlobalConfig) and it turned out to be incompatible due to the
default parsing behavior: an error was returned when unknown fields were
present. In general, for programmatic communications the recommended
behaviour is to silently ignore unknown fields, so that adding new
fields is compatible.

I've also weakened the restriction on accepting new consensus genesis by
external nodes - when the unknown fields are silently dropped, it is
possible that the node will perceive genesis with the same fork number
differently after a binary update (because it receives more fields).