fix(external-node): delete empty unsealed batch on EN initialization (#3125)
## What ❔

This PR reverts #3088 as I have realized it is going to be very hard to
make this fix work by going in that direction. Basically initializing
with an empty unsealed batch causes a lot of issues and the existing
state keeper/external IO flow heavily relies on us having at least one
at the start to initialize correctly. Will leave more context in the
comments.

Feel free to review individual commits to not see revert changelog.

## Why ❔

This bug causes EN to panic

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.