feat(da-clients): add Celestia client (#2983)
## What ❔

This PR adds a Celestia DA client.
The main complexity of this PR comes from our goal to lower the
operational load and not run the Celestia light node (which is a default
way of interacting with Celestia blockchain).

This was done by adapting Astria's Celestia client implementation to our
codebase and removing unneeded logical components.

Note that Celestia's main communication protocol is gRPC, which means we
have to import or maintain the proto definitions. I decided to reuse the
generated `.rs` files from Astria's repo to remove the need to maintain
the `.proto` files in our repo (not the cleanest way, but consider it a
rather temporary solution).
There is a
[celestia-proto](https://github.com/eigerco/lumina/tree/main/proto)
crate that has all the codegen that we need, but they don't generate the
gRPC client definitions, only the types, so we can't use them atm. I
will try to ask the team maintaining it to add such an option, then we
would be able to remove all the codegen from our repo, and simply import
it from celestia-proto.

Example config:
```
da_client:
  celestia:
    api_node_url: http://grpc-mocha.pops.one:9090
    namespace: 000000000000000000000000000000000000ca1de12a5e2d5beb9ba9
    chain_id: mocha-4
    timeout_ms: 10000
``` 
secrets:
```
da:
  celestia:
    private_key: PRIVATE_KEY_WITHOUT_0x_PREFIX
```

## Why ❔

To enable Celestia DA in ZK stack

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.