feat(proof-data-handler): exclude batches without object file in GCS  (#2980)
## What ❔

`/tee/proof_inputs` endpoint no longer returns batches that have no
corresponding object file in Google Cloud Storage for an extended
period.

## Why ❔

TEE's `proof-data-handler` on `mainnet` was flooded with warnings.

Since the recent `mainnet`'s `24.25.0` redeployment, we've been [flooded
with warnings][warnings] for the `proof-data-handler` on `mainnet` (the
warnings are actually _not_ fatal in this context):
```
Failed request with a fatal error

(...)

Blobs for batch numbers 490520 to 490555 not found in the object store. Marked as unpicked.
```
The issue is caused [by the code][code] behind the `/tee/proof_inputs`
[endpoint][endpoint_proof_inputs] (which is equivalent to the
`/proof_generation_data` [endpoint][endpoint_proof_generation_data]) –
it finds the next batch to send to the [requesting][requesting]
`tee-prover` by looking for the first batch that has a corresponding
object in the Google object store. As it skips over batches that don’t
have the objects, [it logs][logging] `Failed request with a fatal error`
for each one (unless the skipped batch was successfully proven, in which
case it doesn’t log the error). This happens with every request the
`tee-prover` sends, which is why we're getting so much noise in the
logs.

One possible solution is to flag the problematic batches as
`permanently_ignored`, like [Thomas did before][Thomas] on `mainnet`.

[warnings]: https://grafana.matterlabs.dev/goto/TjlaXQgHg?orgId=1
[code]:
https://github.com/matter-labs/zksync-era/blob/e162345dfdaaada9d888363a06914132d6ddc7d3/core/node/proof_data_handler/src/tee_request_processor.rs#L35-L79
[endpoint_proof_inputs]:
https://github.com/matter-labs/zksync-era/blob/e162345dfdaaada9d888363a06914132d6ddc7d3/core/node/proof_data_handler/src/lib.rs#L96
[endpoint_proof_generation_data]:
https://github.com/matter-labs/zksync-era/blob/e162345dfdaaada9d888363a06914132d6ddc7d3/core/node/proof_data_handler/src/lib.rs#L67
[requesting]:
https://github.com/matter-labs/zksync-era/blob/e162345dfdaaada9d888363a06914132d6ddc7d3/core/bin/zksync_tee_prover/src/tee_prover.rs#L93
[logging]:
https://github.com/matter-labs/zksync-era/blob/e162345dfdaaada9d888363a06914132d6ddc7d3/core/lib/object_store/src/retries.rs#L56
[Thomas]:
https://matter-labs-workspace.slack.com/archives/C05ANUCGCKV/p1725284962312929

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.