fix(tee): fix race condition in batch locking (#3342)
## What ❔

After [scaling][1] [`zksync-tee-prover`][2] to two instances/replicas on
Azure for `azure-stage2`, `azure-testnet2`, and `azure-mainnet2`, we
started experiencing [duplicated proving for some batches][3].

![logs](https://github.com/user-attachments/assets/39170805-d363-47bf-b275-468694593669)
While this is not an erroneous situation, it is wasteful from a resource
perspective. This was due to a race condition in batch locking. This PR
fixes the issue by adding atomic batch locking.

[1]: https://github.com/matter-labs/gitops-kubernetes/pull/7033/files
[2]:
https://github.com/matter-labs/zksync-era/blob/414275194523814ffd1d433396018ad56a8ad947/core/bin/zksync_tee_prover/src/main.rs
[3]: https://grafana.matterlabs.dev/goto/M1I_Bq7HR?orgId=1

## Why ❔

To fix the bug that only activates after running `zksync-tee-prover` on
multiple instances.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.