feat: Refactor WG DAL (#3560)
## What ❔

Move specified WG DAL methods to separate files. Some of the methods are
unified, so they are left as is, but can be reworked as well in a
follow-up(since it might require another thousand lines of changes).

## Why ❔

To improve readability(the previous file was 2000 lines of code and it's
hard to understand at least something there)

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.