fix(api): Improve estimation for gas_per_pubdata_limit (#3475)
Fixes the `gas_per_pubdata_limit` field calculation. Right now, we
provide the value we get from deriving the existing batch input, which
is most likely lower than limit provided in fee estimation request.
Given that we scale the overall fee, it should be safe to scale the
limit as well, as long as we don't go beyond the limit supplied in the
tx.

Providing the limit as-is without scaling can be very flaky: transaction
risks to be left in the mempool over minor `gas_per_pubdata` changes.