feat(gateway): Requirement to stop L1->L2 transactions before v26 upgrade (#3707)
## What ❔

A way to disable the addition of new L1->L2 transactions completely
before the start of the upgrade.

The chains will be expected to firstly set the new config value to
`true` (and wait until all the older txs are processed) and after the
end of the upgrade return it back to `false` (or remove completely)

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?

This is a breaking change for zkstack_cli functionality for gateway
upgrade as new server version would be required for it to work.

- [x] Yes
- [ ] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.