fix(contract-verifier): contract verifier to return fully matched verification info when available (#3734)
## What ❔

Fix the contract verifier to prioritize fully matched verification info
and return it if available. Only when there is no full match should the
partially matched verification info be returned.

## Why ❔

There is a bug where the contract is not verified, but there are other
verified contracts with the same bytecode and partially matched
bytecode. Despite the presence of a full match, the partially matched
verification info is returned.

## Is this a breaking change?
- [ ] Yes
- [X] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [X] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.