feat(zkos): Implement ZK OS tree manager (#3730)
## What ❔

Implements a tree manager component responsible for updating the ZK OS
Merkle tree based on Postgres data and providing REST API for it (to be
used for witness input generation).

## Why ❔

Necessary to keep the ZK OS Merkle tree synced.

## Is this a breaking change?

- [ ] Yes
- [x] No

## Operational changes

No operational changes; the new component is not tied to the node
binaries.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.